Simplify the dimensional analysis implementation
Simplify the dimensional analysis implementation by tracking only the base units and dimensions of quantities. In addition to this change, add SI and CGS units as well as methods to convert from one system to the other.

Add operators for performing basic arithmetic on quantities. The dimensions returned from the addition and subtraction of quantities are the bitwise-or of the input dimensions. This is because the dimensions of inputs should be the same, and if they are different, they should alter the results' dimensions. The consumer should use VerifyDimensions to validate results.

Remove unused code, rename "Metric" to "SI," and include other minor changes.